Use ProblemDetails middleware
