+++ src/modules/src/Eryph.Modules.AspNetCore/ApiProvider/ApiProviderExtensions.cs
@@ -31,6 +31,8 @@ namespace Eryph.Modules.AspNetCore.ApiProvider
             services.AddOptions<ApiProviderOptions>();
             services.Configure(options);
 
+            services.AddProblemDetails();
+
             services.AddTransient<ICorrelationIdGenerator, CorrelationIdGenerator>();
             //mvcBuilder.AddApplicationPart(typeof(VersionedMetadataController).Assembly);
 
@@ -60,7 +62,7 @@ namespace Eryph.Modules.AspNetCore.ApiProvider
                 {
                     // add a custom operation filter which sets default values
                     options.OperationFilter<SwaggerDefaultValues>();
-                    options.OperationFilter<ApiErrorOperationFilter>();
+                    options.OperationFilter<ProblemDetailsOperationFilter>();
                     options.SchemaFilter<ApiErrorSchemaFilter>();
                     options.SchemaFilter<OperationSchemaFilter>();
                     options.OperationFilter<ListResponseOperationFilter>();
@@ -123,6 +125,9 @@ namespace Eryph.Modules.AspNetCore.ApiProvider
             //var modelBuilder = app.ApplicationServices.GetRequiredService<VersionedODataModelBuilder>();
             var provider = app.ApplicationServices.GetRequiredService<IApiVersionDescriptionProvider>();
 
+            app.UseExceptionHandler();
+            app.UseStatusCodePages();
+
             //var models = modelBuilder.GetEdmModels().ToArray();
             //routing is not supported currently
             //as versioned OData models are currently not supported with endpoint routing.
@@ -156,8 +161,10 @@ namespace Eryph.Modules.AspNetCore.ApiProvider
                 c.SerializeAsV2 = false;
                 c.PreSerializeFilters.Add((swaggerDoc, httpReq) =>
                 {
-                    swaggerDoc.Servers = new List<OpenApiServer>
-                        {new OpenApiServer {Url = module.Path}};
+                    swaggerDoc.Servers =
+                    [
+                        new OpenApiServer { Url = module.Path },
+                    ];
                 });
             });
             app.UseSwaggerUI(
@@ -170,30 +177,6 @@ namespace Eryph.Modules.AspNetCore.ApiProvider
                         options.SwaggerEndpoint($"{module.Path}/swagger/{description.GroupName}/swagger.json",
                             description.GroupName.ToUpperInvariant());
                 });
-
-
-            app.UseExceptionHandler(appBuilder =>
-            {
-                appBuilder.Use(async (context, next) =>
-                {
-                    var env = appBuilder.ApplicationServices.GetRequiredService<IHostEnvironment>();
-                    var error = context.Features[typeof(IExceptionHandlerFeature)] as IExceptionHandlerFeature;
-                    if (error?.Error != null)
-                    {
-                        context.Response.StatusCode = (int)HttpStatusCode.InternalServerError;
-                        context.Response.ContentType = "application/json";
-                        var apiError = ApiError.FromException(error.Error, env.IsDevelopment());
-                       
-                        await context.Response.WriteAsync(JsonSerializer.Serialize(apiError));
-                    }
-
-                    // when no error, do next.
-                    else
-                    {
-                        await next();
-                    }
-                });
-            });
 
             return app;
         }
